财务列表

+++ ezhihui-common/src/main/java/com/ezhihui/www/domain/Teacher.java
@@ -20,5 +20,5 @@ public class Teacher {
 
     private Double salary;
 
-    private Long comment = 0L;
+    private Integer comment = 0;
 }
